Fixed API and added to registration process
Changed the API I'm going to use to MoviesDatabase from RapidAPI.

Added a dislike genre part to registration that will be used to filter out disliked genre's.

Changed the registration to a stack navigator instead of being apart of the drawer navigator.

Changed makeAPICall to accept an options variable for fetch.